refactor address functions into one module
